--- a/src/mbgl/layout/layout.hpp
+++ b/src/mbgl/layout/layout.hpp
@@ -25,9 +25,9 @@ public:
                               bool,
                               const CanonicalTileID&) = 0;
 
-    virtual void prepareSymbols(const GlyphMap&, const GlyphPositions&, const ImageMap&, const ImagePositions&){};
+    virtual void prepareSymbols(const GlyphMap&, const GlyphPositions&, const ImageMap&, const ImagePositions&) {}
 
-    virtual bool hasSymbolInstances() const { return true; };
+    virtual bool hasSymbolInstances() const { return true; }
 
     virtual bool hasDependencies() const = 0;
 };