Is this still worthwhile for negligible performance effect, apparently within the margin of error?

before:

```
| SF zoom in   | 3.62 ms | 6.32 ms |
| SF zoom out  | 1.81 ms | 6.58 ms |
| DC zoom in   | 4.03 ms | 6.43 ms |
| DC zoom out  | 2.23 ms | 7.87 ms |
| AMS zoom in  | 3.40 ms | 5.35 ms |
| AMS zoom out | 1.98 ms | 7.82 ms |
| HEL zoom in  | 3.24 ms | 5.04 ms |
| HEL zoom out | 2.06 ms | 8.33 ms |
| AYA zoom in  | 2.83 ms | 4.62 ms |
| AYA zoom out | 1.78 ms | 5.89 ms |
| BER zoom in  | 3.05 ms | 4.83 ms |
| BER zoom out | 2.10 ms | 8.35 ms |
| BAN zoom in  | 3.18 ms | 5.61 ms |
| BAN zoom out | 2.09 ms | 6.76 ms |
| SHA zoom in  | 1.40 ms | 3.57 ms |
| SHA zoom out | 1.95 ms | 8.12 ms |
Average frame encoding time: 2.55 ms
Average frame rendering time: 6.34 ms
```

```
| SF zoom in   | 3.67 ms | 6.42 ms |
| SF zoom out  | 1.94 ms | 6.66 ms |
| DC zoom in   | 3.80 ms | 6.23 ms |
| DC zoom out  | 2.01 ms | 8.17 ms |
| AMS zoom in  | 3.26 ms | 5.13 ms |
| AMS zoom out | 1.93 ms | 8.14 ms |
| HEL zoom in  | 3.04 ms | 4.98 ms |
| HEL zoom out | 2.04 ms | 8.30 ms |
| AYA zoom in  | 2.73 ms | 4.79 ms |
| AYA zoom out | 1.65 ms | 5.84 ms |
| BER zoom in  | 3.83 ms | 5.52 ms |
| BER zoom out | 2.20 ms | 8.40 ms |
| BAN zoom in  | 3.25 ms | 5.68 ms |
| BAN zoom out | 2.08 ms | 6.89 ms |
| SHA zoom in  | 1.89 ms | 5.18 ms |
| SHA zoom out | 2.19 ms | 8.51 ms |
Average frame encoding time: 2.59 ms
Average frame rendering time: 6.55 ms
```

after:

```
| SF zoom in   | 3.64 ms | 6.26 ms |
| SF zoom out  | 1.90 ms | 6.62 ms |
| DC zoom in   | 3.91 ms | 6.38 ms |
| DC zoom out  | 2.31 ms | 8.14 ms |
| AMS zoom in  | 3.32 ms | 5.11 ms |
| AMS zoom out | 2.08 ms | 8.14 ms |
| HEL zoom in  | 3.29 ms | 5.01 ms |
| HEL zoom out | 2.01 ms | 8.24 ms |
| AYA zoom in  | 2.72 ms | 4.62 ms |
| AYA zoom out | 1.76 ms | 5.80 ms |
| BER zoom in  | 3.76 ms | 5.49 ms |
| BER zoom out | 2.19 ms | 8.31 ms |
| BAN zoom in  | 3.27 ms | 5.69 ms |
| BAN zoom out | 2.03 ms | 6.85 ms |
| SHA zoom in  | 1.43 ms | 4.14 ms |
| SHA zoom out | 1.91 ms | 8.28 ms |
Average frame encoding time: 2.59 ms
Average frame rendering time: 6.44 ms
```

```
| SF zoom in   | 3.61 ms | 6.37 ms |
| SF zoom out  | 1.89 ms | 6.68 ms |
| DC zoom in   | 4.00 ms | 6.47 ms |
| DC zoom out  | 2.36 ms | 8.09 ms |
| AMS zoom in  | 3.40 ms | 5.36 ms |
| AMS zoom out | 1.93 ms | 7.94 ms |
| HEL zoom in  | 3.10 ms | 5.02 ms |
| HEL zoom out | 1.98 ms | 8.32 ms |
| AYA zoom in  | 2.65 ms | 4.62 ms |
| AYA zoom out | 1.66 ms | 5.79 ms |
| BER zoom in  | 3.76 ms | 5.45 ms |
| BER zoom out | 2.13 ms | 8.32 ms |
| BAN zoom in  | 3.13 ms | 5.62 ms |
| BAN zoom out | 2.01 ms | 6.75 ms |
| SHA zoom in  | 1.65 ms | 4.31 ms |
| SHA zoom out | 2.02 ms | 8.28 ms |
Average frame encoding time: 2.58 ms
Average frame rendering time: 6.46 ms
```