@sjg-wdw 2024-02-09
If it's not too complex and doesn't cause any problems, I'd be inclined to keep it.