@TimSylvester 2024-03-06
Updating the metric expectation has resolved the build issue, as long as we're okay with the metrics being dependent on implementation-defined behavior.