@TimSylvester 2024-03-07
The coverage results seem to include a ton of partial hits that don't make much sense.

It seems to consider anywhere that a destructor calls delete (which is safe to call with null) to contain an un-taken branch, but most of those are un-reachable because the thing being deleted was allocated with throwing new.  Similarly, un-failed assertions count as a missing branch, but I don't think we want to test-hit all the assertions even if we can do so without killing the test processes.

<img width="597" alt="image" src="https://github.com/maplibre/maplibre-native/assets/71895881/afd5d2aa-ced7-4a0d-b52a-6a6db097de9b">

<img width="791" alt="image" src="https://github.com/maplibre/maplibre-native/assets/71895881/bc940265-981e-4769-9edf-e55e45f5ef70">

<img width="444" alt="image" src="https://github.com/maplibre/maplibre-native/assets/71895881/5ec86fe2-f310-485a-b8e7-aa67c0ca4209">

The results would be more useful without all that noise, and we could potentially exclude assertions from being compiled in when doing coverage, but it doesn't seem likely there's any way to exclude skipped deletes.